fixed zombie node process when web sockets enabled
Not sure if it's due to WSL2 and the polling, but when the watcher is running, and I interact with the Server, then kill the Server via CTRL+C, one random `node` process hangs around. Sometimes if I waited a really long time it'd eventually die and I'd get the `Server closed` message.

Now I'm keeping track of all open connections to the Server, and killing them all manually. Seems to do the trick for now.