fix: CSS doesn't always load in Dev mode
The Clean plugin runs before CSS assets are extracted, so after a rebuild it'll delete any CSS assets that were previously extracted, and not re-generate the assets if a Dev reverts their changes.
So the manifest ends up pointing to a file that was previously deleted, and not regenerated after the reversion.